integrate Header & HomePage and update Footer
