+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "isuncoin-official",
   "version": "0.1.0+10",
-  "releaseDate": "2024-08-08",
+  "releaseDate": "2024-08-09",
   "private": true,
   "scripts": {
     "dev": "next dev",
