+++ src/components/v1_landing_page/landing_header/landing_header.tsx
@@ -1,11 +1,26 @@
 import Link from 'next/link';
+import useOuterClick from '@/lib/hooks/use_outer_click';
 import { ISUNCOIN_ROUTE_V1 } from '@/constants/url';
 import ThemeToggle from '@/components/toggle/theme_toggle';
-import I18n from '@/components/i18n/i18n';
+
+const languages = [
+  { label: 'English', code: 'en' },
+  { label: '繁體中文', code: 'tw' },
+  { label: '简体中文', code: 'cn' },
+];
 
 const LandingHeader = () => {
   // ToDo: (20240618 - Liz) i18n
   //   const { t }: { t: TranslateFunction } = useTranslation('common');
+
+  const {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    targetRef: dropdownRef,
+    componentVisible: dropdownOpen,
+    setComponentVisible: setDropdownOpen,
+  } = useOuterClick<HTMLDivElement>(false);
+
+  const toggleDropdown = () => setDropdownOpen(!dropdownOpen);
 
   return (
     <header className="fixed inset-x-0 top-0 z-40 flex max-w-full items-center gap-lv-7 bg-surface-neutral-surface-lv1 px-lv-10 py-lv-4 font-barlow shadow-header">
@@ -246,7 +261,40 @@ const LandingHeader = () => {
       </div>
 
       {/* Language Selector Button */}
-      <I18n />
+      <div className="p-12px">
+        <div onClick={toggleDropdown} className="relative">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="32"
+            height="32"
+            fill="none"
+            viewBox="0 0 32 32"
+            className="hover:cursor-pointer hover:text-stroke-brand-primary"
+          >
+            <path
+              // fill="#001840"
+              fill="currentColor"
+              fillRule="evenodd"
+              d="M4.038 13.001a12.36 12.36 0 0 0-.367 3c0 1.035.127 2.04.367 3h5.904a21.4 21.4 0 0 1-.271-3.02c.02-1.004.112-2 .27-2.98H4.039Zm.688-2h5.64a21.403 21.403 0 0 1 3.366-7.124A12.353 12.353 0 0 0 4.726 11ZM16.004 4.2a19.4 19.4 0 0 0-3.564 6.802h7.128A19.4 19.4 0 0 0 16.004 4.2Zm4.033 8.802h-8.066a19.405 19.405 0 0 0-.3 3c.022 1.012.123 2.015.3 3h8.066c.177-.985.278-1.988.3-3a19.405 19.405 0 0 0-.3-3Zm2.029 6a21.41 21.41 0 0 0 .271-3.02 21.409 21.409 0 0 0-.271-2.98h5.904c.24.96.367 1.965.367 3s-.127 2.04-.367 3h-5.904Zm-2.498 2H12.44a19.4 19.4 0 0 0 3.564 6.802A19.4 19.4 0 0 0 19.568 21Zm-5.836 7.125A21.404 21.404 0 0 1 10.365 21H4.726a12.353 12.353 0 0 0 9.006 7.125Zm4.544 0A21.404 21.404 0 0 0 21.643 21h5.639a12.353 12.353 0 0 1-9.006 7.125ZM27.282 11h-5.64a21.403 21.403 0 0 0-3.366-7.124A12.353 12.353 0 0 1 27.282 11ZM1.671 16c0-7.916 6.417-14.333 14.333-14.333S30.337 8.085 30.337 16 23.92 30.334 16.004 30.334 1.67 23.917 1.67 16.001Z"
+              clipRule="evenodd"
+            />
+          </svg>
+          {/* dropmenu */}
+          <ul
+            className={`absolute left-0.5 top-80px flex h-200px w-160px -translate-x-1/2 flex-col items-center justify-evenly rounded-xs border bg-surface-neutral-surface-lv2 px-24px text-center shadow-downDropShadowS ${dropdownOpen ? 'block' : 'hidden'}`}
+          >
+            {languages.map((lang: { label: string; code: string }) => (
+              <li
+                key={lang.code}
+                className="cursor-pointer text-base font-medium text-button-text-secondary"
+                // onClick={() => changeLanguage(lang.code)} // ToDo: (20240618 - Liz) i18n
+              >
+                {lang.label}
+              </li>
+            ))}
+          </ul>
+        </div>
+      </div>
 
       {/* Theme Toggle (Day/Night mode) Button */}
       <ThemeToggle />
