+++ src/pages/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import Head from 'next/head';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import { ILocale } from '@/interfaces/locale';
 import Header from '@/components/common/Header';
 import HomePageBody from '@/components/home_page/home_page_body';
 import Footer from '@/components/common/Footer';
@@ -12,16 +14,24 @@ const HomePage = () => {
         <title>iSunCoin</title>
       </Head>
 
-      {/* Navbar */}
+      {/* Info:(20240808 - Julian) Navbar */}
       <Header />
 
-      {/* Body */}
+      {/* Info:(20240808 - Julian) Body */}
       <HomePageBody />
 
-      {/* Footer */}
+      {/* Info:(20240808 - Julian) Footer */}
       <Footer />
     </>
   );
 };
+
+const getStaticPropsFunction = async ({ locale }: ILocale) => ({
+  props: {
+    ...(await serverSideTranslations(locale, ['common'])),
+  },
+});
+
+export const getStaticProps = getStaticPropsFunction;
 
 export default HomePage;
